+++ src/Sequence/Defer.php
@@ -507,12 +507,16 @@ final class Defer implements Implementation
             (static function(\Iterator $values, callable $function): \Generator {
                 /** @var callable(T, T): int $sorter */
                 $sorter = $function;
+                $loaded = [];
 
-                /** @var list<T> */
-                $values = \iterator_to_array($values);
-                \usort($values, $sorter);
-
+                /** @var T $value */
                 foreach ($values as $value) {
+                    $loaded[] = $value;
+                }
+
+                \usort($loaded, $sorter);
+
+                foreach ($loaded as $value) {
                     yield $value;
                 }
             })($this->values, $function),
@@ -554,9 +558,16 @@ final class Defer implements Implementation
         /** @psalm-suppress ImpureFunctionCall */
         return new self(
             (static function(\Iterator $values): \Generator {
-                $values = \iterator_to_array($values);
+                $reversed = [];
 
-                yield from \array_reverse($values);
+                /** @var T $value */
+                foreach ($values as $value) {
+                    \array_unshift($reversed, $value);
+                }
+
+                foreach ($reversed as $value) {
+                    yield $value;
+                }
             })($this->values),
         );
     }
@@ -768,7 +779,12 @@ final class Defer implements Implementation
      */
     private function load(): Implementation
     {
-        /** @psalm-suppress ImpureFunctionCall */
-        return new Primitive(\array_values(\iterator_to_array($this->values)));
+        $values = [];
+
+        foreach ($this->values as $value) {
+            $values[] = $value;
+        }
+
+        return new Primitive($values);
     }
 }