+++ src/Sequence/Lazy.php
@@ -538,12 +538,17 @@ final class Lazy implements Implementation
 
         return new self(
             static function() use ($values, $function): \Generator {
+                $loaded = [];
+
                 // bypass the registering of cleanup function as we iterate over
                 // the whole generator
-                $values = \iterator_to_array($values(self::bypassCleanup()));
-                \usort($values, $function);
+                foreach ($values(self::bypassCleanup()) as $value) {
+                    $loaded[] = $value;
+                }
 
-                foreach ($values as $value) {
+                \usort($loaded, $function);
+
+                foreach ($loaded as $value) {
                     yield $value;
                 }
             },
@@ -586,11 +591,17 @@ final class Lazy implements Implementation
 
         return new self(
             static function() use ($values): \Generator {
+                $reversed = [];
+
                 // bypass the registering of cleanup function as we iterate over
                 // the whole generator
-                $values = \iterator_to_array($values(self::bypassCleanup()));
+                foreach ($values(self::bypassCleanup()) as $value) {
+                    \array_unshift($reversed, $value);
+                }
 
-                yield from \array_reverse($values);
+                foreach ($reversed as $value) {
+                    yield $value;
+                }
             },
         );
     }
@@ -831,8 +842,13 @@ final class Lazy implements Implementation
      */
     private function load(): Implementation
     {
-        /** @psalm-suppress ImpureFunctionCall */
-        return new Primitive(\array_values(\iterator_to_array($this->iterator())));
+        $values = [];
+
+        foreach ($this->iterator() as $value) {
+            $values[] = $value;
+        }
+
+        return new Primitive($values);
     }
 
     /**