+++ src/Sequence/Primitive.php
@@ -126,8 +126,13 @@ final class Primitive implements Implementation
      */
     public function equals(Implementation $sequence): bool
     {
-        /** @psalm-suppress ImpureFunctionCall */
-        return $this->values === \iterator_to_array($sequence->iterator());
+        $other = [];
+
+        foreach ($sequence->iterator() as $value) {
+            $other[] = $value;
+        }
+
+        return $this->values === $other;
     }
 
     /**
@@ -350,11 +355,13 @@ final class Primitive implements Implementation
      */
     public function append(Implementation $sequence): self
     {
-        /** @psalm-suppress ImpureFunctionCall */
-        return new self(\array_merge(
-            $this->values,
-            \iterator_to_array($sequence->iterator()),
-        ));
+        $other = [];
+
+        foreach ($sequence->iterator() as $value) {
+            $other[] = $value;
+        }
+
+        return new self(\array_merge($this->values, $other));
     }
 
     /**
@@ -519,9 +526,13 @@ final class Primitive implements Implementation
         $aggregate = new Aggregate($this->iterator());
         /** @psalm-suppress MixedArgument */
         $values = $aggregate(static fn($a, $b) => $exfiltrate($map($a, $b))->iterator());
+        $aggregated = [];
 
-        /** @psalm-suppress ImpureFunctionCall */
-        return new self(\array_values(\iterator_to_array($values)));
+        foreach ($values as $value) {
+            $aggregated[] = $value;
+        }
+
+        return new self($aggregated);
     }
 
     /**