+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - A lazy `Sequence::slice()` no longer loads the whole underlying `Generator`
 - `Innmind\Immutable\Set::matches()`, `Innmind\Immutable\Sequence::matches()` and `Innmind\Immutable\Map::matches()` no longer iterates over all elements when one value doesn't match the predicate
+- When using `yield from` in the `Generator` passed to `Sequence::lazy()` values may be lost on certain operations
 
 ## 4.14.0 - 2023-04-29